Refactors datasets creation into standalone file.
